% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{all}
\alias{all}
\title{Returns TRUE iff all elements match the predicate, otherwise FALSE.}
\usage{
all(pred, xs)
}
\arguments{
\item{pred}{predicate function}

\item{xs}{list}
}
\description{
\code{all :: (a -> Bool) -> [a] -> Bool}
}
