% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prelude.r
\name{and}
\alias{and}
\title{Combine all logical values of a list with a conjuntion. Returns @TRUE@ on an empty list.}
\usage{
and(xs)
}
\arguments{
\item{xs}{list of Bools}
}
\description{
\code{and :: [Bool] -> Bool}
}
