% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{any}
\alias{any}
\title{Returns TRUE if any element matches the predicate, otherwise FALSE.}
\usage{
any(pred, xs, skipNAs = TRUE)
}
\arguments{
\item{pred}{predicate function}

\item{xs}{list}
}
\description{
\code{any :: (a -> Bool) -> [a] -> Bool}
}
