% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{concatMap}
\alias{concatMap}
\title{concatMap is \code{concat} \verb{\%.\%} \code{map}. Here the list can be used as monoid, neglecting some values,
while concat collects the accepted values.}
\usage{
concatMap(f, xs)
}
\arguments{
\item{f}{function to apply}

\item{xs}{list}
}
\description{
\code{concatMap :: (a -> [b]) -> [a] -> [b]}
}
