% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{filter}
\alias{filter}
\title{Filters a list by predicate function.}
\usage{
filter(f, xs)
}
\arguments{
\item{f}{predicate function}

\item{xs}{list}
}
\description{
\code{filter :: (a -> Bool) -> [a] -> [a]}
}
