% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{foldl}
\alias{foldl}
\title{Fold over a list applying a function to each element and accumulating the result.}
\usage{
foldl(f, acc, xs)
}
\arguments{
\item{f}{accumulator function}

\item{acc}{accumulator}

\item{xs}{list}
}
\description{
\code{foldl :: (b -> a -> b) -> b -> [a] -> b}
}
