% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Functor.r
\name{Functor}
\alias{Functor}
\title{Functor class.}
\description{
Functor class.

Functor class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Functor$new()}}
\item \href{#method-fmap}{\code{Functor$fmap()}}
\item \href{#method-clone}{\code{Functor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Functor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmap"></a>}}
\if{latex}{\out{\hypertarget{method-fmap}{}}}
\subsection{Method \code{fmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Functor$fmap(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{function

\code{pure :: (a -> b) -> f a (=self) -> f b}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Functor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
