% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prelude.r
\name{\%.\%}
\alias{\%.\%}
\title{Standard math. function composition: \code{(f \%.\% g)(x) == f(g(x))}}
\usage{
f \%.\% g
}
\arguments{
\item{f}{function}

\item{g}{function}
}
\description{
\code{\%.\% :: (b -> c) -> (a -> b) -> (a -> c)}
}
