% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prelude.r
\name{\%notElem\%}
\alias{\%notElem\%}
\title{Is not element of list.}
\usage{
x \%notElem\% xs
}
\arguments{
\item{a}{element to check}

\item{[a]}{list}
}
\description{
\code{\%notElem\% :: a -> [a] -> Bool}
}
