% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{head}
\alias{head}
\title{Returns the first element of a list. Unsafe! I.e. expects a non-empty list, or will throw an error.}
\usage{
head(xs)
}
\arguments{
\item{xs}{list}
}
\description{
\code{head :: [a] -> a}
}
