% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{intercalate}
\alias{intercalate}
\title{'intercalate' @xs xss@ is equivalent to @('concat' ('intersperse' xs xss))@. It inserts the list}
\usage{
intercalate(xs, xss)
}
\arguments{
\item{xs}{element}

\item{xss}{list of elements}
}
\description{
'intercalate' @xs xss@ is equivalent to @('concat' ('intersperse' xs xss))@. It inserts the list
}
