% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{intersperse}
\alias{intersperse}
\title{O(n). The intersperse function takes an element and a list and `intersperses' that element
between the elements of the list.}
\usage{
intersperse(x, xs)
}
\arguments{
\item{xs}{list of elements}

\item{a}{element}
}
\description{
\code{intersperse :: a -> [a] -> [a]}
}
