% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{last}
\alias{last}
\title{Returns the last element of a list. Unsafe! I.e. expects a non-empty list, or will throw an error.}
\usage{
last(xs)
}
\arguments{
\item{xs}{list}
}
\description{
\code{last :: [a] -> a}
}
