% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{map}
\alias{map}
\title{Map with corrected parameter order. You should only use pure functions! Use @mapM_@ in case of
impure functions.}
\usage{
map(f, xs)
}
\arguments{
\item{f}{function to apply}

\item{xs}{list}
}
\description{
\code{map :: (a -> b) -> [a] -> [b]}
}
