% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{mapM_}
\alias{mapM_}
\title{This is @map@, but with indication that IO is happening. I.e. the program state can change and a
reader must have a closer look what the function does.}
\usage{
mapM_(f, xs)
}
\arguments{
\item{f}{function to apply}

\item{xs}{list}
}
\description{
\code{mapM_ :: (a -> IO b) -> [a] -> IO ()}
}
