% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Maybe.r
\name{Maybe}
\alias{Maybe}
\title{Maybe monad implementation.}
\description{
Maybe monad implementation.

Maybe monad implementation.
}
\section{Super classes}{
\code{\link[rhaskell:Functor]{rhaskell::Functor}} -> \code{\link[rhaskell:Applicative]{rhaskell::Applicative}} -> \code{\link[rhaskell:Monad]{rhaskell::Monad}} -> \code{Maybe}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Maybe$new()}}
\item \href{#method-fmap}{\code{Maybe$fmap()}}
\item \href{#method-pure}{\code{Maybe$pure()}}
\item \href{#method-apply}{\code{Maybe$apply()}}
\item \href{#method-bind}{\code{Maybe$bind()}}
\item \href{#method-fromJust}{\code{Maybe$fromJust()}}
\item \href{#method-fromMaybe}{\code{Maybe$fromMaybe()}}
\item \href{#method-maybe}{\code{Maybe$maybe()}}
\item \href{#method-clone}{\code{Maybe$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$new(mValue = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mValue}}{value.

Return the value. Fails if the encapsulated value is NULL! Better use \code{fromMaybe}!

\code{isJust :: Maybe a (=self) -> a}
Unwrap the value with a default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmap"></a>}}
\if{latex}{\out{\hypertarget{method-fmap}{}}}
\subsection{Method \code{fmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$fmap(fun)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pure"></a>}}
\if{latex}{\out{\hypertarget{method-pure}{}}}
\subsection{Method \code{pure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$pure(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$apply(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bind"></a>}}
\if{latex}{\out{\hypertarget{method-bind}{}}}
\subsection{Method \code{bind()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$bind(fun)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJust"></a>}}
\if{latex}{\out{\hypertarget{method-fromJust}{}}}
\subsection{Method \code{fromJust()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$fromJust()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromMaybe"></a>}}
\if{latex}{\out{\hypertarget{method-fromMaybe}{}}}
\subsection{Method \code{fromMaybe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$fromMaybe(def)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{def}}{default value

\code{fromMaybe :: a -> Maybe a (=self) -> a}
Unwrap a maybe with default value and a function to apply.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maybe"></a>}}
\if{latex}{\out{\hypertarget{method-maybe}{}}}
\subsection{Method \code{maybe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$maybe(def, f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{def}}{default value}

\item{\code{f}}{function to apply in case of \code{Just()} value

\code{maybe :: a -> (a -> b) -> Maybe a (=self) -> a}
Check if a \code{Maybe} is of type \code{Just}.

\code{isJust :: Maybe a (=self) -> Bool}
Check if a \code{Maybe} is of type \code{Nothing}.

\code{isJust :: Maybe a (=self) -> Bool}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
