% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prelude.r
\name{or}
\alias{or}
\title{Combine all logical values of a list with a disjuntion. Return @FALSE@ on an empty list.}
\usage{
or(xs)
}
\arguments{
\item{xs}{list of Bools}
}
\description{
\code{or :: [Bool] -> Bool}
}
