% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Ord.r
\name{pGt}
\alias{pGt}
\title{Partially applied greater, returns a function with > x, where x is the first parameter.}
\usage{
pGt(x)
}
\arguments{
\item{x}{parameter x in y > x}
}
\description{
\code{pGt :: a -> b -> Bool}
}
\details{
E.g. pGt(3)(4) = TRUE
E.g. pGt(3)(2) = FALSE
}
