% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Ord.r
\name{pLeq}
\alias{pLeq}
\title{Partially applied less or equal, returns a function with <= x, where x is the first parameter.}
\usage{
pLeq(x)
}
\arguments{
\item{x}{parameter x in y <= x}
}
\description{
\code{pLeq :: a -> b -> Bool}
}
\details{
E.g. pLeq(3)(4) = FALSE
E.g. pLeq(3)(2) = TRUE
}
