% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Ord.r
\name{pNIdentical}
\alias{pNIdentical}
\title{Partially applied @not \%.\% identical@, returns a function, where the first parameter is set.}
\usage{
pNIdentical(x)
}
\arguments{
\item{x}{first parameter value in @!identical(x, y)@}
}
\description{
\code{pIdentical :: a -> b -> Bool}
}
