% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{replicate}
\alias{replicate}
\title{Create a list of @nr@ elements of @x@.}
\usage{
replicate(nr, x)
}
\arguments{
\item{nr}{int}

\item{x}{any type}
}
\description{
\code{replicate :: Int -> a -> [a]}
}
