% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prelude.r
\name{subtract}
\alias{subtract}
\title{Subtraction.}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{Numeric value}

\item{y}{Numeric value}
}
\description{
\code{subtract :: Num a => a -> a -> a}
}
\details{
subtract(3, 4) = -1
}
