% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Tuple.r
\name{uncurry}
\alias{uncurry}
\title{Function to apply a tuple to a function. Makes a function taking a tuple instead of two
parameters.}
\usage{
uncurry(f)
}
\arguments{
\item{f}{function taking 2 parameters}
}
\description{
\code{uncurry :: (a -> b -> c) -> ((a, b) -> c)}
}
