% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{zip}
\alias{zip}
\title{Zip two list into a list of tuples.}
\usage{
zip(xs, ys)
}
\arguments{
\item{xs}{list. One list to zip.}

\item{ys}{list. Another list to zip.}
}
\description{
\code{zip :: [a] -> [b] -> [(a, b)]}
}
