% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{zipWith}
\alias{zipWith}
\title{Zip two lists together by a specific function.}
\usage{
zipWith(f, xs, ys)
}
\arguments{
\item{f}{function. A function that should take two arguments, iterating over each element of both
lists.}

\item{xs}{list. One list to zip.}

\item{ys}{list. Another list to zip.}
}
\description{
\code{zipWith :: (a -> b -> c) -> [a] -> [b] -> [c]}
}
