% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{zipWith3}
\alias{zipWith3}
\title{Zip two lists together by a specific function.}
\usage{
zipWith3(f, xs, ys, zs)
}
\arguments{
\item{f}{function. A function that should take two arguments, iterating over each element of both
lists.}

\item{xs}{list.}

\item{ys}{list.}

\item{zs}{list.}
}
\description{
\code{concat :: [[a]] -> [a]}
}
