% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Ord.r
\name{pGeq}
\alias{pGeq}
\title{Partially applied greater or equal, returns a function with >= x, where x is the first parameter.}
\usage{
pGeq(x)
}
\arguments{
\item{x}{parameter x in y >= x}
}
\description{
\code{pGeq :: a -> b -> Bool}
}
\details{
E.g. pGeq(3)(4) = TRUE
E.g. pGeq(3)(2) = FALSE
}
