% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.List.r
\name{zipWith3}
\alias{zipWith3}
\title{Zip two lists together by a specific function.}
\usage{
zipWith3(f, xs, ys, zs)
}
\arguments{
\item{f}{function. A funciton that should take two arguments, iterating over each element of both
lists.}

\item{xs}{list. One list to zip.}

\item{ys}{list. Antoher list to zip.}
}
\description{
Zip two lists together by a specific function.
}
